requirement.txt
django==1.9
elasticsearch-dsl==7.0.0
mysqlclient==1.4.4
django-redis==4.10.0
djangorestframework==3.10.3
